lestarch: fixing several bugs with `fprime-util`

+++ Fw/Python/src/fprime/util/build_helper.py
@@ -91,7 +91,7 @@ def validate(parsed):
         # Generation validation
         elif parsed.build_dir is None:
             build_inst_name = fprime.fbuild.cmake.CMakeHandler.CMAKE_DEFAULT_BUILD_NAME.format(parsed.platform)
-            parsed.build_dir = os.path.join(os.getcwd(), build_inst_name)
+            parsed.build_dir = os.path.join(parsed.path, build_inst_name)
     except fprime.fbuild.cmake.CMakeProjectException as exc:
         print("[ERROR] {}".format(exc))
         sys.exit(1)
@@ -131,7 +131,7 @@ def validate(parsed):
             toolchains = list(filter(os.path.exists, toolchains_paths))
             if not toolchains:
                 print("[ERROR] Toolchain file {} does not exist at any of {}"
-                      .format(toolchains + ".cmake", ", ".join(list(toolchains_paths))))
+                      .format(toolchain + ".cmake", ", ".join(list(toolchains_paths))))
                 sys.exit(-1)
             print("[INFO] Using toolchain file {} for platform {}".format(toolchains[0], parsed.platform))
             cmake_args.update({"CMAKE_TOOLCHAIN_FILE": toolchains[0]})
@@ -218,8 +218,9 @@ def utility_entry(args=sys.argv[1:]):
     Main interface to F prime utility.
     :return: return code of the function.
     """
-    parsed, cmake_args, make_args, automatic_build_dir = parse_args(args)
+    parsed = None
     try:
+        parsed, cmake_args, make_args, automatic_build_dir = parse_args(args)
         if parsed.command == "hash-to-file":
             suffix = UT_SUFFIX if parsed.unittest else ""
             lines = fprime.fbuild.builder().find_hashed_file(parsed.build_dir + suffix, parsed.hash)
@@ -257,9 +258,13 @@ def utility_entry(args=sys.argv[1:]):
             fprime.fbuild.builder().execute_known_target(action["target"], parsed.build_dir + action["build-suffix"],
                                                          parsed.path, cmake_args, make_args,
                                                          action.get("top-target", False))
+    except fprime.fbuild.cmake.CMakeExecutionException as exexc:
+        stderr = exexc.get_errors()
+        print("[ERROR] {}.{}".format(exexc,"\n{}".format(stderr) if stderr else ""), file=sys.stderr)
+        return 1
     except fprime.fbuild.cmake.CMakeException as exc:
         print("[ERROR] {}".format(exc), file=sys.stderr)
-        if parsed.command == "generate" and automatic_build_dir:
+        if parsed is not None and parsed.command == "generate" and automatic_build_dir:
             print("[INFO] Cleaning automatic build directory at: {}".format(parsed.build_dir))
             shutil.rmtree(parsed.build_dir, ignore_errors=True)
             print("[INFO] Cleaning automatic unit-test build directory at: {}".format(parsed.build_dir + UT_SUFFIX))
